git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@47398 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/simplification/ExpressionSimplificationVisitor.java
@@ -337,8 +337,8 @@ public class ExpressionSimplificationVisitor
 
     // enum constant
     if (decl instanceof CEnumerator) {
-      final long v = ((CEnumerator) decl).getValue();
-      return new CIntegerLiteralExpression(expr.getFileLocation(), type, BigInteger.valueOf(v));
+      final BigInteger v = ((CEnumerator) decl).getValue();
+      return new CIntegerLiteralExpression(expr.getFileLocation(), type, v);
     }
 
     // const variable, inline initializer
